Group outside of render, extract to separate files
